requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy>=2.0.32
jinja2==3.1.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
pyjwt==2.8.0
python-dotenv==1.0.1
slowapi==0.1.9

